+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 ### Added
 - Keep last playback to restore it even if player destroyed
 - Properly handle alarm creating and destroying player
-- Add button to play playlist firectly from app
+- Add button to play playlist directly from app
+- Handle alarm snoozed
 
 ## [1.0.0] - 2022-01-29
 ### Added